+++ src/mqtt/dc_mqtt.c
@@ -1,4 +1,4 @@
-#include "dc_mqtt.h"
+#include <discusFW/mqtt/dc_mqtt.h>
 
 typedef struct
 {
